+++ app/controllers/application_controller.rb
@@ -2,6 +2,7 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception, prepend: true
 
   include SessionCurrentUser
+  include ValidRequest
   include Pundit
 
   def require_http_auth
@@ -46,24 +47,6 @@ class ApplicationController < ActionController::Base
     params[:i] == "i"
   end
   helper_method :internal_navigation?
-
-  def valid_request_origin?
-    # This manually does what it was supposed to do on its own.
-    # We were getting this issue:
-    # HTTP Origin header (https://dev.to) didn't match request.base_url (http://dev.to)
-    # Not sure why, but once we work it out, we can delete this method.
-    # We are at least secure for now.
-    return if Rails.env.test?
-
-    if request.referer.present?
-      request.referer.start_with?(ApplicationConfig["APP_PROTOCOL"].to_s + ApplicationConfig["APP_DOMAIN"].to_s)
-    else
-      logger.info "**REQUEST ORIGIN CHECK** #{request.origin}"
-      raise InvalidAuthenticityToken, NULL_ORIGIN_MESSAGE if request.origin == "null"
-
-      request.origin.nil? || request.origin.gsub("https", "http") == request.base_url.gsub("https", "http")
-    end
-  end
 
   def set_no_cache_header
     response.headers["Cache-Control"] = "no-cache, no-store"
