+++ app/controllers/display_ad_events_controller.rb
@@ -1,5 +1,7 @@
-class DisplayAdEventsController < ApplicationController
+class DisplayAdEventsController < ApplicationMetalController
+  include ActionController::Head
   # No policy needed. All views are for all users
+
   def create
     # Only tracking for logged in users at the moment
     display_ad_event_create_params = display_ad_event_params.merge(user_id: session_current_user_id)
@@ -25,6 +27,6 @@ class DisplayAdEventsController < ApplicationController
   end
 
   def display_ad_event_params
-    params.require(:display_ad_event).permit(%i[context_type category display_ad_id])
+    params[:display_ad_event].slice(:context_type, :category, :display_ad_id)
   end
 end
