+++ app/controllers/html_variant_successes_controller.rb
@@ -1,4 +1,6 @@
-class HtmlVariantSuccessesController < ApplicationController
+class HtmlVariantSuccessesController < ApplicationMetalController
+  include ActionController::Head
+
   def create
     HtmlVariantSuccessCreateJob.perform_later(html_variant_id: params[:html_variant_id], article_id: params[:article_id])
     head :ok
