+++ app/controllers/html_variant_trials_controller.rb
@@ -1,4 +1,6 @@
-class HtmlVariantTrialsController < ApplicationController
+class HtmlVariantTrialsController < ApplicationMetalController
+  include ActionController::Head
+
   def create
     HtmlVariantTrialCreateJob.perform_later(html_variant_id: params[:html_variant_id], article_id: params[:article_id])
     head :ok
