+++ spec/factories/articles.rb
@@ -15,7 +15,7 @@ FactoryBot.define do
       with_collection { nil }
     end
     association :user, factory: :user, strategy: :create
-    description   { Faker::Hipster.paragraph(sentence_count: 1)[0..100] }
+    description { Faker::Hipster.paragraph(sentence_count: 1)[0..100] }
     main_image    { Faker::Avatar.image }
     language { "en" }
     experience_level_rating { rand(4..6) }
